fix: include project root in package.path so requires work
When the CWD is not the project root, `require("lib.Demojify")` would
fail because lua cant find the file.

We add the data-dir argument when needed, setting it to the lib
folder, which contains an init.lua file. This init.lua file prepends the
project root to the package.path which allows `require("lib.Demojify")`
to succeed.