fix: default version in description (#68)
## Problem

The details for description are:

```text
Project description used in title (if empty, uses neovim version and current date)
```

However this currently does not work, instead the version section of the
description is left empty.

## Solution

The reason this occurs is `vimversion` is set to empty by default:

```bash
"--metadata=vimversion:${VIM_VERSION:-""}"
```

However the check compares the value to `nil`.

Added a helper `empty` function that checks if the value is `nil` or if
the length is 0. Used when checking `vim_version` & `project_description`.